backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sungminpark <devmin0714@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.8"
Django = "4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
django-jazzmin = "^2.6.0"
shortuuid = "^1.0.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
